feat: enable it for safari (iOS, iPadOS, OSX)
After update iPhone and iPad to OS 15 I recognized that iDevices won't
work. Had to change copy mechanism to navigator.clipboard.writeText.
Problem is that function doesn't work nested in functions. Old copy
button must start file download but clipbard.writeText had to wait for
it. So installed a new button only for put-to-clipboard function.
Old copy button tries to click this new button if file is downloaded.
That is working on chrome, opera and firefox. On safari you must click
this new button by yourself.

Signed-off-by: Conrad Zelck <git@simpel.cc>