feat: color buttons green/red on request status
The copy button is disabled by default because there is nothing to copy.

If request for file is ok the pull button will turn green for 2s. If
auto clicking the copy button is successful it will turn green for 2s
too. The copy button will be enabled for 7s.

If the request is 404 the pull button will turn red for 2s.

Signed-off-by: Conrad Zelck <git@simpel.cc>